Adapt to libfirm 9b70628.
This removes the manual changes to the automatically generated RTA
bindings.  The capability to set the RTA callback is only used by
bytecode2firm, which uses the C API.  The Java API does not have any
users.

If a Java frontend needs to set this callback, refer to
https://jna.java.net/nonav/javadoc/overview-summary.html#callbacks
as well as jFirm's src/firm/Firm.java and src/firm/DebugInfo.java.

Also, this commit makes a trivial adaption to libfirm 1f55516.